requirements.txt
discord.py
asyncpg
asyncpg-stubs
typing-extensions
jishaku